k-food-helpdesk/server/requirements.txt
fastapi
uvicorn[standard]
openai
psycopg2-binary
python-dotenv
